+++ tools/cosmovisor/cmd/cosmovisor/version_test.go
@@ -8,7 +8,6 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"cosmossdk.io/log"
-	"cosmossdk.io/tools/cosmovisor"
 )
 
 func TestVersionCommand_Error(t *testing.T) {
@@ -24,5 +23,5 @@ func TestVersionCommand_Error(t *testing.T) {
 	ctx := context.WithValue(context.Background(), log.ContextKey, logger)
 
 	require.Error(t, rootCmd.ExecuteContext(ctx))
-	require.Contains(t, out.String(), cosmovisor.ErrEmptyConfigENV.Error())
+	require.Contains(t, out.String(), "DAEMON_NAME is not set")
 }