+++ tools/cosmovisor/flags.go
@@ -7,5 +7,4 @@ const (
 	FlagForce             = "force"
 	FlagUpgradeHeight     = "upgrade-height"
 	FlagConfig            = "config"
-	FlagExportConfig      = "export-config"
 )